backend/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-noise-cancellation
requests
mem0ai
duckduckgo-search
langchain_community
python-dotenv
pydantic-ai-slim[openai,mcp]
livekit-agents[tavus]~=1.0
fastapi
uvicorn[standard]
livekit